+++ mirx/src/image/mod.rs
@@ -51,7 +51,7 @@ mod plan;
 mod surface;
 mod view;
 
-pub use borrowed::{ImageSource, SurfacePlanes, SurfaceView};
+pub use borrowed::{ImageSource, SurfaceCopyError, SurfacePlanes, SurfaceView};
 pub use color::{
     ChromaSiting, ColorDescription, ColorDescriptionError, ColorMatrix, ColorPrimaries, ColorRange,
     TransferFunction,